backend/requirements.txt
# Core Django
Django==4.2.11
psycopg2-binary==2.9.9
python-decouple==3.8

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1

# Authentication & Security
django-allauth==0.61.1
django-otp==1.4.1
qrcode[pil]==7.4.2
django-axes==6.1.1
django-csp==3.8
django-permissions-policy==4.28.0
django-ipware==7.0.1
cryptography==42.0.5
bleach==6.1.0
defusedxml==0.7.1
django-dbbackup==4.3.0
# django-session-security==2.6.3  # Désactivé - incompatible avec Django 4.2 (utilise django.conf.urls.url supprimé)

# File Storage
django-storages==1.14.2
boto3==1.34.51
Pillow==10.2.0

# Async Tasks
django-q2==1.6.2

# Notifications
sendgrid==6.11.0
twilio==9.0.4

# Payments
stripe==8.7.0

# PDF Generation (GRATUIT)
reportlab==4.0.9
weasyprint==60.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
hashids==1.3.1  # Obfuscation des IDs dans les URLs

# Testing & Quality
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
isort==5.13.2
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring (GRATUIT)
sentry-sdk==1.40.6

# Excel/CSV Export (GRATUIT)
openpyxl==3.1.2
xlsxwriter==3.2.0
